Extracted code converting a reflist to a string to a separate class.
